com.unity.memoryprofiler@1.1.5 ## [1.1.5] - 2025-02-14
### Added
- Managed fields referencing native allocations are now parsed as references. References show up in the References To tab for the managed objects holding the references, as well as in the Referenced By tab for native allocations under `Native > Unity Subsystems > UnsafeUtility` in the All Of Memory table. Note that this is not an exhaustive list of all references to these allocations, as stack variables could reference them as well but are not part of the snapshot. Thus, a native allocation without references is not necessarily leaked.
- Native Allocations (e.g. those listed in `Native > Unity Subsystems > UnsafeUtility` in the All Of Memory table) are now named based on the field and object information of the first managed reference to them, making them easier to parse than address values and stand out from the list as having reference information.

### Fixed
- The Boehm GC treats any pointer sized field as potential managed pointer. So far the Memory Profiler only examined reference fields to find referenced objects (outside of those directly held by static fields or GCHandles). The Memory Profiler does not yet follow this logic as it would end up finding invalid Managed objects, but it now finds references from unsafe pointer fields to Managed Objects that are also held by a GC Handle.
- The Managed Fields table now shows any pointer sized fields that might be holding managed objects as referencing these managed objects (Note: if they are not referenced otherwise, they won't sho up in the other tables).

### Changed
- The References To table now shows field information for references held by managed fields

+++ Editor/Containers/MetaDataHelpers.cs
@@ -94,7 +94,7 @@ namespace Unity.MemoryProfiler.Editor.Containers
             }
         }
 
-        internal static bool GenerateMetaDataString(CachedSnapshot cs, int nativeObjectIndex, out List<(string, string)> metaDataStrings)
+        internal static bool GenerateMetaDataString(CachedSnapshot cs, long nativeObjectIndex, out List<(string, string)> metaDataStrings)
         {
             if (!cs.HasNativeObjectMetaData || nativeObjectIndex == CachedSnapshot.NativeObjectEntriesCache.InvalidObjectIndex)
             {
